# lib/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: jambonz REST API
  description: jambonz REST API
  contact:
    email: daveh@drachtio.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: /v1
    description: development server
paths:
...
  /Accounts/{AccountSid}/SpeechCredentials:
    parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: add a speech credential
      operationId: createSpeechCredential
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechCredential'
      responses:
        201:
          description: speech credential successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulAdd'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      summary: retrieve all speech credentials for an account
      operationId: listSpeechCredentials
      responses:
        200:
          description: retrieve speech credentials for a specified account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpeechCredential'
        404:
          description: account not found
  /Accounts/{AccountSid}/SpeechCredentials/{SpeechCredentialSid}:
    parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: SpeechCredentialSid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: get a specific speech credential
      operationId: getSpeechCredential
      responses:
        200:
          description: retrieve speech credentials for a specified account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeechCredential'
        404:
          description: credential not found
    put:
      summary: update a speech credential
      operationId: updateSpeechCredential
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechCredentialUpdate'
      responses:
        204:
          description: credential successfully deleted
        404:
          description: credential not found
        422:
          description: credential not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    delete:
      summary: delete a speech credential
      operationId: deleteSpeechCredential
      responses:
        204:
          description: credential successfully deleted
        404:
          description: credential not found
  /Accounts/{AccountSid}/SpeechCredentials/{SpeechCredentialSid}/test:
    get:
      summary: test a speech credential
      operationId: testSpeechCredential
      parameters:
        - name: AccountSid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: SpeechCredentialSid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: credential test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  tts:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: 
                            - success 
                            - fail
                            - not tested
                      reason:
                        type: string
                  stt:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: 
                            - success 
                            - fail
                      reason:
                        type: string
        404:
          description: credential not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
  /Accounts/{AccountSid}/Calls:
    post:
      summary: create a call
      operationId: createCall
      parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
                - from
                - to
              type: object
              properties:
                application_sid:
                  type: string
                  format: uuid
                  description: The application to use to control this call.  Either applicationSid or url is required.
                call_hook:
                  $ref: '#/components/schemas/Webhook'
                call_status_hook:
                  $ref: '#/components/schemas/Webhook'
                  example: {"url": "https://acme.com/status", "method": "POST"}
                from:
                  type: string
                  description: The calling party number
                  example: "16172375089"
                timeout:
                  type: integer
                  description: the number of seconds to wait for call to be answered.  Defaults to 60.
                  example: 30
                tag:
                  type: object
                  description: initial set of customer-supplied metadata to associate with the call (see jambonz 'tag' verb)
                  example: {"callCount": 10}
                to:
                  $ref: '#/components/schemas/Target'
                  description: destination for call
      responses:
        201:
          description: call successfully created
          content:
            application/json:
              schema:
                required:
                  - sid
                properties:
                  sid:
                    type: string
                    format: uuid
                    example: 2531329f-fb09-4ef7-887e-84e648214436
        400:
          description: bad request
    get:
      summary: list calls
      operationId: listCalls
      parameters:
      - name: AccountSid
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: list of calls for a specified account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Call'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
...
    SuccessfulAdd:
      type: object
      required:
        - sid
      properties:
        sid:
          type: string
      example:
        sid: 9d26a637-1679-471f-8da8-7150266e1254
    GeneralError:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
      example:
        msg: specific error detail will be provided here
...
    Webhook:
      type: object
      properties:
        webhook_sid:
          type: string
          format: uuid
        url:
          type: string
          format: url
        method:
          type: string
          enum:
            - get
            - post
        username:
          type: string
        password:
          type: string
      required:
        - url
      example: {"url": "https://acme.com", "method": "POST"}
...
    Call:
      type: object
      properties:
        account_sid:
          type: string
          format: uuid
        application_sid:
          type: string
          format: uuid
        call_id:
          type: string
        call_sid:
          type: string
          format: uuid
        call_status:
          type: string
          enum:
            - trying
            - ringing
            - alerting
            - in-progress
            - completed
            - busy
            - no-answer
            - failed
            - queued
        caller_name:
          type: string
        direction:
          type: string
          enum:
            - inbound
            - outbound 
        duration:
          type: integer
        from:
          type: string
        originating_sip_trunk_name:
          type: string
        parent_call_sid:
          type: string
          format: uuid
        service_url:
          type: string
        sip_status:
          type: integer
        to:
          type: string
      required:
        - account_sid
        - call_id
        - call_sid
        - call_status
        - direction
        - from
        - service_url
        - sip_status
        - to
    Target:
      properties:
        type:
          type: string
          enum:
            - phone
            - sip
            - user
        number:
          type: string
        sipUri:
          type: string
        name:
          type: string
        auth:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
      required:
        - type
      example: {"type": "phone", "number": "+16172375080"}
...
    SpeechCredential:
      properties:
        speech_credential_sid:
          type: string
          format: uuid
        account_sid:
          type: string
          format: uuid
        vendor:
          type: string
          enum:
            - google
            - aws
        service_key:
          type: string
        access_key_id:
          type: string
        secret_access_key:
          type: string
        last_used:
          type: string
          format: date-time
        last_tested:
          type: string
          format: date-time
        use_for_tts:
          type: boolean
        use_for_stt:
          type: boolean
        tts_tested_ok:
          type: boolean
        stt_tested_ok:
          type: boolean
    SpeechCredentialUpdate:
      properties:
        use_for_tts:
          type: boolean
        use_for_stt:
          type: boolean
...
security:
  - bearerAuth: []